ShortListとLonglistクラスを作成
